--- src/main/kotlin/com/hrothwell/anime/commands/Search.kt
@@ -1,58 +1,0 @@
-package com.hrothwell.anime.commands
-
-import com.github.ajalt.clikt.core.CliktCommand
-import com.github.ajalt.clikt.parameters.options.default
-import com.github.ajalt.clikt.parameters.options.option
-import com.github.ajalt.clikt.parameters.options.required
-import com.hrothwell.anime.client.MalClient
-import com.hrothwell.anime.util.AnimeUtil
-
-class Search : CliktCommand(
-  help = """
-    Search MAL for an anime/manga
-  """.trimIndent()
-) {
-
-  private val keywords by option(
-    "-q", "--query", help = """
-    search terms to use, surround multiple words in quotes: "Tokyo Ghoul" or 'Tokyo Ghoul'
-  """.trimIndent()
-  ).required()
-
-  private val limit by option(
-    "-l", "--limit", help = """
-    limit of items to return from the search. MAL limits to 100
-  """.trimIndent()
-  ).default("1")
-
-  override fun run() {
-    try {
-      if (AnimeUtil.isAnime) {
-        val animeList = MalClient.getAnimeList(keywords, limit.toInt())
-        echo(animeList.map { it.title })
-        if (animeList.size == 1) {
-          AnimeUtil.openAnime(animeList.first())
-        }
-      } else {
-        val mangaList = MalClient.getMangaList(keywords, limit.toInt())
-        echo(mangaList.map { it.title })
-        if (mangaList.size == 1) {
-          AnimeUtil.openManga(mangaList.first())
-        }
-      }
-
-    } catch (t: Throwable) {
-      echoError(
-        """
-        error searching anime / manga. Message: ${t.message}
-        
-        cause: ${t.cause}
-      """.trimIndent()
-      )
-    }
-  }
-
-  private fun echoError(msg: String) {
-    echo("${AnimeUtil.RED} $msg ${AnimeUtil.RESET}", err = true)
-  }
-}
