--- src/main/kotlin/com/hrothwell/anime/commands/Suggest.kt
@@ -1,40 +1,0 @@
-package com.hrothwell.anime.commands
-
-import com.github.ajalt.clikt.core.CliktCommand
-import com.github.ajalt.clikt.parameters.options.default
-import com.github.ajalt.clikt.parameters.options.option
-import com.hrothwell.anime.client.MalClient
-import com.hrothwell.anime.util.AnimeUtil
-
-class Suggest : CliktCommand(
-  help = "Get anime suggestions! Manga suggestions not yet supported by MAL"
-) {
-
-  val limit by option(
-    "-l", "--limit", help = """
-    limit on the amount of anime to suggest. Default 1, max 100
-  """.trimIndent()
-  )
-    .default("1")
-
-  override fun run() {
-    try {
-      val animeList = MalClient.getSuggestedAnime(limit.toInt())
-      echo("Try: ${animeList.map { it.title }}")
-      if (animeList.size == 1) {
-        AnimeUtil.openAnime(animeList.first())
-      }
-    } catch (t: Throwable) {
-      echoError(
-        """
-        unable to suggest anime. Message: ${t.message}
-        cause: ${t.cause}
-      """.trimIndent()
-      )
-    }
-  }
-
-  private fun echoError(msg: String) {
-    echo("${AnimeUtil.RED} $msg ${AnimeUtil.RESET}", err = true)
-  }
-}
