--- src/main/kotlin/com/hrothwell/anime/domain/mal/AnimeAiringStatus.kt
@@ -1,16 +1,0 @@
-package com.hrothwell.anime.domain.mal
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-enum class AnimeAiringStatus {
-  @SerialName("finished_airing")
-  FINISHED_AIRING,
-
-  @SerialName("currently_airing")
-  CURRENTLY_AIRING,
-
-  @SerialName("not_yet_aired")
-  NOT_YET_AIRED
-}
