--- src/main/kotlin/com/hrothwell/anime/domain/mal/AnimeNode.kt
@@ -1,18 +1,0 @@
-package com.hrothwell.anime.domain.mal
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class AnimeNode(
-  val id: Long,
-  val title: String,
-  val status: AnimeAiringStatus? = null,
-  val mean: Float? = null,
-  val rank: Int? = null,
-  val popularity: Int? = null,
-  @SerialName("num_episodes")
-  val numEpisodes: Int? = null,
-  @SerialName("average_episode_duration")
-  val averageEpisodeDurationSeconds: Int? = null,
-)
