--- src/main/kotlin/com/hrothwell/anime/domain/mal/MalGenericListResponse.kt
@@ -1,9 +1,0 @@
-package com.hrothwell.anime.domain.mal
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class MalGenericListResponse<T>(
-  val data: List<T>,
-  // TODO paging
-)
