--- src/main/kotlin/com/hrothwell/anime/domain/mal/MalOAuthResponse.kt
@@ -1,16 +1,0 @@
-package com.hrothwell.anime.domain.mal
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class MalOAuthResponse(
-  @SerialName("token_type")
-  val tokenType: String,
-  @SerialName("expires_in")
-  val expiresIn: Long,
-  @SerialName("access_token")
-  val accessToken: String,
-  @SerialName("refresh_token")
-  val refreshToken: String
-)
