--- src/main/kotlin/com/hrothwell/anime/exception/LoginException.kt
@@ -1,6 +1,0 @@
-package com.hrothwell.anime.exception
-
-data class LoginException(
-  val msg: String,
-  val rootCause: Throwable? = null
-) : Throwable(msg)
