--- src/main/kotlin/com/hrothwell/anime/exception/MALResponseException.kt
@@ -1,5 +1,0 @@
-package com.hrothwell.anime.exception
-
-data class MALResponseException(
-  val msg: String
-) : Throwable(msg)
