--- src/main/kotlin/com/hrothwell/anime/exception/OAuthCallException.kt
@@ -1,6 +1,0 @@
-package com.hrothwell.anime.exception
-
-data class OAuthCallException(
-  val msg: String,
-  val rootCause: Throwable?
-) : Throwable(msg, rootCause)
