--- src/main/kotlin/com/hrothwell/anime/exception/UserSecretsException.kt
@@ -1,6 +1,0 @@
-package com.hrothwell.anime.exception
-
-data class UserSecretsException(
-  val msg: String,
-  val rootCause: Throwable
-) : Throwable(msg, rootCause)
