+++ src/main/kotlin/com/hrothwell/mal/domain/response/AnimeData.kt
@@ -1,0 +1,11 @@
+package com.hrothwell.mal.domain.response
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class AnimeData(
+  val node: AnimeNode,
+  @SerialName("list_status")
+  val listStatus: AnimeUserListStatus? = null
+)
