+++ src/main/kotlin/com/hrothwell/mal/domain/response/MalGenericListResponse.kt
@@ -1,0 +1,9 @@
+package com.hrothwell.mal.domain.response
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class MalGenericListResponse<T>(
+  val data: List<T>,
+  // TODO paging
+)
