+++ src/main/kotlin/com/hrothwell/mal/domain/response/MangaData.kt
@@ -1,0 +1,11 @@
+package com.hrothwell.mal.domain.response
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class MangaData(
+  val node: MangaNode,
+  @SerialName("list_status")
+  val listStatus: MangaUserListStatus? = null
+)
