+++ src/main/kotlin/com/hrothwell/mal/exception/LoginException.kt
@@ -1,0 +1,6 @@
+package com.hrothwell.mal.exception
+
+data class LoginException(
+  val msg: String,
+  val rootCause: Throwable? = null
+) : Throwable(msg)
