+++ src/main/kotlin/com/hrothwell/mal/exception/MALResponseException.kt
@@ -1,0 +1,5 @@
+package com.hrothwell.mal.exception
+
+data class MALResponseException(
+  val msg: String
+) : Throwable(msg)
