+++ src/main/kotlin/com/hrothwell/mal/exception/OAuthCallException.kt
@@ -1,0 +1,6 @@
+package com.hrothwell.mal.exception
+
+data class OAuthCallException(
+  val msg: String,
+  val rootCause: Throwable?
+) : Throwable(msg, rootCause)
