+++ src/main/kotlin/com/hrothwell/mal/exception/UserSecretsException.kt
@@ -1,0 +1,6 @@
+package com.hrothwell.mal.exception
+
+data class UserSecretsException(
+  val msg: String,
+  val rootCause: Throwable
+) : Throwable(msg, rootCause)
